+++ packages/oxlint-plugin-react-doctor/src/plugin/rules/a11y/no-target-blank-without-rel.ts
@@ -1,32 +1,12 @@
 import { defineRule } from "../../utils/define-rule.js";
-import type { EsTreeNode } from "../../utils/es-tree-node.js";
 import type { EsTreeNodeOfType } from "../../utils/es-tree-node-of-type.js";
 import { findJsxAttribute } from "../../utils/find-jsx-attribute.js";
-import { getJsxPropStringValue } from "../../utils/get-jsx-prop-string-value.js";
+import { getStringLiteralAttributeValue } from "../../utils/get-string-literal-attribute-value.js";
 import { hasJsxSpreadAttribute } from "../../utils/has-jsx-spread-attribute.js";
 import { isNodeOfType } from "../../utils/is-node-of-type.js";
 
 const MESSAGE =
   '`<a target="_blank">` without `rel="noopener"` lets the opened page script your tab via `window.opener` (reverse tabnabbing). Add `rel="noopener noreferrer"`.';
-
-// Resolve an attribute to its string value from a plain literal (`rel="x"`) or
-// a string literal inside an expression container (`rel={'x'}`). Returns null
-// for a dynamic value (`rel={rel}`) or boolean shorthand. Used for both
-// `target` and `rel` so `rel={'noopener'}` is judged, not silently skipped.
-const getStringLiteralAttributeValue = (
-  attribute: EsTreeNodeOfType<"JSXAttribute">,
-): string | null => {
-  const stringValue = getJsxPropStringValue(attribute);
-  if (stringValue !== null) return stringValue;
-  const value = attribute.value as EsTreeNode | null;
-  if (value && isNodeOfType(value, "JSXExpressionContainer")) {
-    const expression = value.expression;
-    if (isNodeOfType(expression, "Literal") && typeof expression.value === "string") {
-      return expression.value;
-    }
-  }
-  return null;
-};
 
 export const noTargetBlankWithoutRel = defineRule({
   id: "no-target-blank-without-rel",
